requirements.txt
streamlit
selenium
beautifulsoup4
html2text
pandas
pydantic
tiktoken
streamlit-tags
google-generativeai